Revert "Finalize instanced rendering with backend abstraction"
This reverts commit c198c536370f1804e0651a86ac65d6a7a8ec7f9d.
