+++ nomad-realms/src/main/java/engine/nengen/DrawBatch.java
@@ -1,29 +1,152 @@
 package engine.nengen;
 
+import static engine.common.colour.Colour.normalizedA;
+import static engine.common.colour.Colour.normalizedB;
+import static engine.common.colour.Colour.normalizedG;
+import static engine.common.colour.Colour.normalizedR;
+import static org.lwjgl.opengl.GL15.GL_STREAM_DRAW;
+
+import java.util.ArrayList;
+import java.util.List;
+
 import engine.common.math.Matrix4f;
+import engine.visuals.lwjgl.GLContext;
+import engine.visuals.lwjgl.render.ShaderProgram;
+import engine.visuals.lwjgl.render.VertexArrayObject;
+import engine.visuals.lwjgl.render.VertexBufferObject;
 
 /**
  * A DrawBatch collects instance data (transforms and colors) and renders them in a single instanced draw call.
- * This interface provides a backend-agnostic way for the game logic to perform batch rendering.
+ * It reuses OpenGL resources (VAO and VBOs) across frames to minimize overhead.
+ * <p>
+ * Use {@link #add(Matrix4f, int)} to add instances, then {@link #draw()} to render them.
+ * Call {@link #clear()} before each frame's data collection.
  */
-public interface DrawBatch {
+public class DrawBatch {
 
-	/**
-	 * Adds an instance to the batch.
-	 *
-	 * @param transform the transformation matrix for the instance
-	 * @param color     the color for the instance
-	 */
-	void add(Matrix4f transform, int color);
+	private final List<Matrix4f> transforms = new ArrayList<>();
+	private final List<Integer> colors = new ArrayList<>();
 
-	/**
-	 * Clears all instances from the batch. Should be called before each frame's data collection.
-	 */
-	void clear();
+	private VertexArrayObject vao;
+	private ShaderProgram shaderProgram;
+	private GLContext glContext;
 
-	/**
-	 * Renders all instances currently in the batch.
-	 */
-	void draw();
+	private VertexArrayObject instancedVao;
+	private VertexBufferObject tVbo;
+	private VertexBufferObject cVbo;
+
+	private int lastCount = -1;
+
+	public DrawBatch() {
+	}
+
+	public DrawBatch vao(VertexArrayObject vao) {
+		this.vao = vao;
+		return this;
+	}
+
+	public DrawBatch shaderProgram(ShaderProgram shaderProgram) {
+		this.shaderProgram = shaderProgram;
+		return this;
+	}
+
+	public DrawBatch glContext(GLContext glContext) {
+		this.glContext = glContext;
+		return this;
+	}
+
+	public void add(Matrix4f transform, int color) {
+		transforms.add(transform);
+		colors.add(color);
+	}
+
+	public void clear() {
+		transforms.clear();
+		colors.clear();
+	}
+
+	public void draw() {
+		if (transforms.isEmpty()) {
+			return;
+		}
+
+		int count = transforms.size();
+		float[] transformData = new float[count * 16];
+		float[] colorData = new float[count * 4];
+
+		for (int i = 0; i < count; i++) {
+			transforms.get(i).store(transformData, i * 16);
+			int color = colors.get(i);
+			colorData[i * 4 + 0] = normalizedR(color);
+			colorData[i * 4 + 1] = normalizedG(color);
+			colorData[i * 4 + 2] = normalizedB(color);
+			colorData[i * 4 + 3] = normalizedA(color);
+		}
+
+		if (instancedVao == null) {
+			tVbo = new VertexBufferObject()
+					.index(2)
+					.dimensions(4)
+					.data(transformData)
+					.stride(16 * Float.BYTES)
+					.offset(0)
+					.usage(GL_STREAM_DRAW)
+					.load();
+			tVbo.divisor(1);
+
+			VertexBufferObject tVbo2 = new VertexBufferObject()
+					.index(3)
+					.dimensions(4)
+					.stride(16 * Float.BYTES)
+					.offset(4 * Float.BYTES)
+					.loadConfig(tVbo.id());
+			tVbo2.divisor(1);
+
+			VertexBufferObject tVbo3 = new VertexBufferObject()
+					.index(4)
+					.dimensions(4)
+					.stride(16 * Float.BYTES)
+					.offset(8 * Float.BYTES)
+					.loadConfig(tVbo.id());
+			tVbo3.divisor(1);
+
+			VertexBufferObject tVbo4 = new VertexBufferObject()
+					.index(5)
+					.dimensions(4)
+					.stride(16 * Float.BYTES)
+					.offset(12 * Float.BYTES)
+					.loadConfig(tVbo.id());
+			tVbo4.divisor(1);
+
+			cVbo = new VertexBufferObject()
+					.index(6)
+					.dimensions(4)
+					.data(colorData)
+					.usage(GL_STREAM_DRAW)
+					.load();
+			cVbo.divisor(1);
+
+			instancedVao = new VertexArrayObject()
+					.ebo(vao.ebo())
+					.vbos(vao.vbos().toArray(new VertexBufferObject[0]))
+					.vbos(tVbo, tVbo2, tVbo3, tVbo4, cVbo)
+					.load();
+			lastCount = count;
+		} else {
+			tVbo.data(transformData);
+			cVbo.data(colorData);
+			if (count > lastCount) {
+				tVbo.reallocate();
+				cVbo.reallocate();
+				lastCount = count;
+			} else {
+				tVbo.updateData();
+				cVbo.updateData();
+			}
+		}
+
+		shaderProgram.use(glContext);
+		instancedVao.drawInstanced(glContext, count);
+	}
 
 }
