--- nomad-realms/src/main/java/engine/visuals/lwjgl/render/LwjglDrawBatch.java
@@ -1,143 +1,0 @@
-package engine.visuals.lwjgl.render;
-
-import static engine.common.colour.Colour.normalizedA;
-import static engine.common.colour.Colour.normalizedB;
-import static engine.common.colour.Colour.normalizedG;
-import static engine.common.colour.Colour.normalizedR;
-import static org.lwjgl.opengl.GL15.GL_STREAM_DRAW;
-
-import java.util.ArrayList;
-import java.util.List;
-
-import engine.common.math.Matrix4f;
-import engine.nengen.DrawBatch;
-import engine.visuals.lwjgl.GLContext;
-
-public class LwjglDrawBatch implements DrawBatch {
-
-	private final List<Matrix4f> transforms = new ArrayList<>();
-	private final List<Integer> colors = new ArrayList<>();
-
-	private VertexArrayObject vao;
-	private ShaderProgram shaderProgram;
-	private GLContext glContext;
-
-	private VertexArrayObject instancedVao;
-	private VertexBufferObject tVbo;
-	private VertexBufferObject cVbo;
-
-	private int lastCount = -1;
-
-	public LwjglDrawBatch vao(VertexArrayObject vao) {
-		this.vao = vao;
-		return this;
-	}
-
-	public LwjglDrawBatch shaderProgram(ShaderProgram shaderProgram) {
-		this.shaderProgram = shaderProgram;
-		return this;
-	}
-
-	public LwjglDrawBatch glContext(GLContext glContext) {
-		this.glContext = glContext;
-		return this;
-	}
-
-	@Override
-	public void add(Matrix4f transform, int color) {
-		transforms.add(transform);
-		colors.add(color);
-	}
-
-	@Override
-	public void clear() {
-		transforms.clear();
-		colors.clear();
-	}
-
-	@Override
-	public void draw() {
-		if (transforms.isEmpty()) {
-			return;
-		}
-
-		int count = transforms.size();
-		float[] transformData = new float[count * 16];
-		float[] colorData = new float[count * 4];
-
-		for (int i = 0; i < count; i++) {
-			transforms.get(i).store(transformData, i * 16);
-			int color = colors.get(i);
-			colorData[i * 4 + 0] = normalizedR(color);
-			colorData[i * 4 + 1] = normalizedG(color);
-			colorData[i * 4 + 2] = normalizedB(color);
-			colorData[i * 4 + 3] = normalizedA(color);
-		}
-
-		if (instancedVao == null) {
-			tVbo = new VertexBufferObject()
-					.index(2)
-					.dimensions(4)
-					.data(transformData)
-					.stride(16 * Float.BYTES)
-					.offset(0)
-					.usage(GL_STREAM_DRAW)
-					.load();
-			tVbo.divisor(1);
-
-			VertexBufferObject tVbo2 = new VertexBufferObject()
-					.index(3)
-					.dimensions(4)
-					.stride(16 * Float.BYTES)
-					.offset(4 * Float.BYTES)
-					.loadConfig(tVbo.id());
-			tVbo2.divisor(1);
-
-			VertexBufferObject tVbo3 = new VertexBufferObject()
-					.index(4)
-					.dimensions(4)
-					.stride(16 * Float.BYTES)
-					.offset(8 * Float.BYTES)
-					.loadConfig(tVbo.id());
-			tVbo3.divisor(1);
-
-			VertexBufferObject tVbo4 = new VertexBufferObject()
-					.index(5)
-					.dimensions(4)
-					.stride(16 * Float.BYTES)
-					.offset(12 * Float.BYTES)
-					.loadConfig(tVbo.id());
-			tVbo4.divisor(1);
-
-			cVbo = new VertexBufferObject()
-					.index(6)
-					.dimensions(4)
-					.data(colorData)
-					.usage(GL_STREAM_DRAW)
-					.load();
-			cVbo.divisor(1);
-
-			instancedVao = new VertexArrayObject()
-					.ebo(vao.ebo())
-					.vbos(vao.vbos().toArray(new VertexBufferObject[0]))
-					.vbos(tVbo, tVbo2, tVbo3, tVbo4, cVbo)
-					.load();
-			lastCount = count;
-		} else {
-			tVbo.data(transformData);
-			cVbo.data(colorData);
-			if (count > lastCount) {
-				tVbo.reallocate();
-				cVbo.reallocate();
-				lastCount = count;
-			} else {
-				tVbo.updateData();
-				cVbo.updateData();
-			}
-		}
-
-		shaderProgram.use(glContext);
-		instancedVao.drawInstanced(glContext, count);
-	}
-
-}
